backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0
psycopg[binary]>=3.1.0
alembic>=1.13.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
httpx>=0.27.0
fpdf2>=2.8.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
